fix the texture image path, add a local simple HTTP server bootstrap script
Local file access by a browser is prohibited, so texture loads are subject to
this limitation. I thought it had to do with the relative position of the
loading javascript file, so I had messed up the path. Restored it to the
proper path "image/disc.png", and added a simple HTTP server script to
simplify starting a local HTTP server for testing.