Adding visualization of the Reduced Dependency Graph.
Tested a couple of graph visualization libraries: sigma.js, d3.js, and visjs.
sigma.js was causing me some grief as I couldn't quite figure out how all the
events work, and it didn't have an out of the box direct graph renderer.
Similarly, with d3.js, amount of work to get a simple graph going was too
much, however, d3.js is very attractive and at one point will be brought in to
provide dashboard and performance statistics. visjs has a digraph renderer out
of the box, and supports the DOT language, which I was familiar with.